First perlin

+++ Source/Math/Distance.cs
@@ -3,15 +3,15 @@ namespace Wang.MathW
 {
     static class Distance
     {
-        public static float EuclidianDistance(int firstX, int firstY, int secondX, int secondY)
+        public static float EuclidianDistance(float firstX, float firstY, float secondX, float secondY)
         {
-            int diffX = firstX - secondX;
-            int diffY = firstY - secondY;
+            float diffX = firstX - secondX;
+            float diffY = firstY - secondY;
 
             return MathF.Sqrt(diffX * diffX + diffY * diffY);
         }
 
-        public static int ManhatamDistance(int firstX, int firstY, int secondX, int secondY)
+        public static float ManhatamDistance(float firstX, float firstY, float secondX, float secondY)
         { 
             return Math.Abs(firstX - secondX) + System.Math.Abs(firstY - secondY);
         }
